Separate console as a plugin
It's not strictly part of the core.
