+++ gaphor/plugins/console/__init__.py
@@ -1,0 +1,1 @@
+from gaphor.plugins.console.consolewindow import ConsoleWindow
