+++ gaphor/plugins/console/console.py
