+++ gaphor/plugins/console/consolewindow.py
@@ -10,7 +10,7 @@ from gaphor.action import action
 from gaphor.misc import get_config_dir
 from gaphor.abc import ActionProvider
 from gaphor.ui.abc import UIComponent
-from gaphor.misc.console import GTKInterpreterConsole
+from gaphor.plugins.console.console import GTKInterpreterConsole
 
 log = logging.getLogger(__name__)
