+++ setup.py
@@ -107,7 +107,7 @@ setup(
             "namespace = gaphor.ui.namespace:Namespace",
             "toolbox = gaphor.ui.toolbox:Toolbox",
             "diagrams = gaphor.ui.mainwindow:Diagrams",
-            "consolewindow = gaphor.ui.consolewindow:ConsoleWindow",
+            "consolewindow = gaphor.plugins.console:ConsoleWindow",
             "elementeditor = gaphor.ui.elementeditor:ElementEditor",
             "diagram_export = gaphor.plugins.diagramexport:DiagramExport",
             "xmi_export = gaphor.plugins.xmiexport:XMIExport",
