+++ win-installer/gaphor-script.py
@@ -2,6 +2,7 @@ if __name__ == "__main__":
     import gaphor
     from gaphor import core
     from gaphor.UML.elementfactory import ElementFactory
+    from gaphor.plugins.console import ConsoleWindow
     from gaphor.plugins.diagramexport import DiagramExport
     from gaphor.plugins.xmiexport import XMIExport
     from gaphor.services.componentregistry import ComponentRegistry
@@ -11,7 +12,6 @@ if __name__ == "__main__":
     from gaphor.services.properties import Properties
     from gaphor.services.sanitizerservice import SanitizerService
     from gaphor.services.undomanager import UndoManager
-    from gaphor.ui.consolewindow import ConsoleWindow
     from gaphor.ui.elementeditor import ElementEditor
     from gaphor.ui.filemanager import FileManager
     from gaphor.ui.mainwindow import Diagrams
